fix(progress-animation): update story knobs, rename file
